+++ CHANGELOG.md
@@ -1,8 +1,23 @@
+## 0.0.9 (2018-04-21)
+
+Feature
+
+- Extend Api On `SocketServer`
+  - `createClient(id)`
+
+- Extend Api On `SocketClient`
+  - `set/getTag`
+  - `addChannel`
+  - `getChannel`
+  - `isInChannel(channel:String/String[])` *TODO*
+  - `set/getAlias`
+  - `addExtendData(data:Object)`
+
 ## 0.0.8 (2018-04-20)
 
 Feature
 
-- Add SocketClient
+- Add SocketClient Class
 
 
 ## 0.0.7 (2018-04-20)
@@ -17,6 +32,7 @@ Feature
   - `bindTimeoutEvent(fn)` 绑定客户端超时事件的函数
   - `bindCloseEvent(fn)`   绑定客户端关闭事件的函数
   - `bindErrorEvent(fn)`   绑定客户端异常事件的函数
+  - `bindReceiveEvent(fn)`  绑定接收到的数据处理函数
   - `broadcast(message, channel)`  广播消息,渠道可选
   - `send(message, clientId)` 给指定的客户端发送消息
 
@@ -31,5 +47,6 @@ Feature
 
 Bugs
 
-- Change `localhost` to `0.0.0.0`
+- Change `localhost` to `0.0.0.0` 
+  Fix Remote Mode Cant Reach The Server .
 - Add output Offline Client Id